Rename files, make conventional dir structure
