+++ bin/read_receipt_file.rb
@@ -2,7 +2,7 @@
 $:.push "#{File.dirname(__FILE__)}/../lib"
 
 require 'debug'
-require 'sr500receiptfile'
+require 'sr500_receipt_file'
 
 obj = SR500RecipeFile.new("./src2/EJ250405.TXT")
 obj.readlines
