+++ lib/jp_encoding_file.rb
