+++ lib/sr500_constants.rb
