+++ lib/sr500_receipt_file.rb
@@ -1,9 +1,9 @@
-require 'jpencodingfile'
-require 'sr500order'
-require 'sr500constants'
+require 'jp_encoding_file'
+require 'sr500_order'
+require 'sr500_constants'
 require 'debug'
 require_relative 'receipt_parser'
-require_relative 'SR500Order'
+require_relative 'sr500_order'
 
 # A class for reading recepit files of Casi SR-500 POS register.
 class SR500ReceiptFile < JPEncodingFile
